Apply changes from upstream
https://github.com/KDE/konsole/commit/06b13420bf46e8cfc95a949a3ff838dd87a8364b
https://github.com/KDE/konsole/commit/be60d9151d3afde1e873d68be9b498cb9713081e
https://github.com/KDE/konsole/commit/6cc32028e1adfe7d4833e4c68296ae9148e474b9
https://github.com/KDE/konsole/commit/bed79d9fdf3eb96ecc8cd8a672c61cdd6cdca212
https://github.com/KDE/konsole/commit/e40230a710041fd207b94480047804178ae844b6
https://github.com/KDE/konsole/commit/6b990498e7df854a0dc5deef3150ef4dd90fd443
https://github.com/KDE/konsole/commit/3805cb9d9ddf51e19dff2a76bb9b24dc6346fa7c

+++ src/ColorScheme.cpp
@@ -469,7 +469,7 @@ bool ColorSchemeWallpaper::draw(QPainter& painter, const QRect rect, qreal opaci
     if (!_picture || _picture->isNull())
         return false;
 
-    if (qFuzzyCompare(1.0, opacity)) {
+    if (qFuzzyCompare(qreal(1.0), opacity)) {
         painter.drawTiledPixmap(rect, *_picture, rect.topLeft());
         return true;
     }
